refactor: update directory and namespace
